Frontend: Move save/load operations to api/contributions module
