+++ vue-app/src/components/ContributionModal.vue
@@ -54,10 +54,8 @@ import { Prop } from 'vue-property-decorator'
 import { BigNumber, Contract, FixedNumber, Signer } from 'ethers'
 import { Keypair, PubKey, Message } from 'maci-domainobjs'
 
-import { Contributor } from '@/api/contributions'
+import { saveContributorData } from '@/api/contributions'
 import { RoundInfo } from '@/api/round'
-import { storage } from '@/api/storage'
-import { User } from '@/api/user'
 import Transaction from '@/components/Transaction.vue'
 import { LOAD_ROUND_INFO } from '@/store/action-types'
 import { SET_CONTRIBUTOR, SET_CONTRIBUTION } from '@/store/mutation-types'
@@ -65,26 +63,6 @@ import { waitForTransaction, getEventArg } from '@/utils/contracts'
 import { createMessage } from '@/utils/maci'
 
 import { FundingRound, ERC20, MACI } from '@/api/abi'
-
-const CONTRIBUTOR_INFO_STORAGE_KEY = 'contributor-info'
-
-function saveContributorInfo(
-  fundingRoundAddress: string,
-  user: User,
-  contributor: Contributor,
-) {
-  const serializedData = JSON.stringify({
-    privateKey: contributor.keypair.privKey.serialize(),
-    stateIndex: contributor.stateIndex,
-  })
-  storage.setItem(
-    user.walletAddress,
-    user.encryptionKey,
-    fundingRoundAddress,
-    CONTRIBUTOR_INFO_STORAGE_KEY,
-    serializedData,
-  )
-}
 
 @Component({
   components: {
@@ -176,10 +154,10 @@ export default class ContributionModal extends Vue {
       keypair: contributorKeypair,
       stateIndex: stateIndex.toNumber(),
     }
-    // Save contributor info to storage
-    saveContributorInfo(
-      fundingRoundAddress,
+    // Save contributor data to storage
+    saveContributorData(
       this.$store.state.currentUser,
+      fundingRoundAddress,
       contributor,
     )
     // Set contribution and update round info
